# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Đăng ký khách hàng (CUSTOMER)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Username đã tồn tại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login-unified": {
      "post": {
        "operationId": "AuthController_loginUnified",
        "summary": "Đăng nhập chung (admin hoặc khách)",
        "description": "Body: `{ account, password }`. Admin: account = email. Khách: account = username. Set cookie refreshToken.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnifiedLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Đăng nhập khách (username + password)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Làm mới access token",
        "description": "Cookie `refreshToken` (credentials: include). Không cần Bearer.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "AuthController_session",
        "summary": "Phiên hiện tại (CUSTOMER hoặc ADMIN)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/api/v1/leads": {
      "post": {
        "operationId": "LeadsPublicController_create",
        "summary": "Gửi đơn đăng ký / tư vấn",
        "description": "Optional Bearer khách → gắn customerId. Rate limit theo IP/SĐT.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadCreateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Public — Leads"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/leads/history": {
      "get": {
        "operationId": "LeadsPublicController_historyByPhone",
        "summary": "Tra cứu lịch sử đăng ký theo SĐT (không đăng nhập)",
        "parameters": [
          {
            "name": "phone",
            "required": true,
            "in": "query",
            "description": "Số điện thoại đã dùng khi đăng ký",
            "schema": {
              "example": "0912345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadHistoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Public — Leads"
        ]
      }
    },
    "/api/v1/leads/history/{id}": {
      "get": {
        "operationId": "LeadsPublicController_historyDetail",
        "summary": "Chi tiết một đơn (SĐT phải khớp)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MongoDB ObjectId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": true,
            "in": "query",
            "description": "Số điện thoại đã dùng khi đăng ký",
            "schema": {
              "example": "0912345678",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadPublicItemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Public — Leads"
        ]
      }
    },
...
    "/api/v1/admin/auth/login": {
      "post": {
        "operationId": "AdminAuthController_login",
        "summary": "Đăng nhập quản trị (email + password)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin — Auth"
        ]
      }
    },
...
    "/api/v1/admin/leads": {
      "get": {
        "operationId": "AdminLeadsController_list",
        "summary": "Danh sách lead (filter + phân trang)",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "NEW",
                "CONTACTED",
                "CONVERTED",
                "CANCELLED"
              ],
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "ISO date",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "ISO date",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin — Leads"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/admin/leads/{id}": {
      "patch": {
        "operationId": "AdminLeadsController_patch",
        "summary": "Cập nhật trạng thái / ghi chú lead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchAdminLeadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin — Leads"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "FPT Telecom Landing API",
    "description": "REST API cho landing page & admin FPT Telecom.\n\n**Base path:** `/api/v1`\n\n**Xác thực:**\n- Access token: header `Authorization: Bearer <accessToken>`\n- Refresh token: cookie HttpOnly `refreshToken` (credentials: include)\n\n**FE public (không bắt buộc login):** packages, leads, faqs, navigation, package-quiz, banners\n\n**Rate limit:** leads history, lead create theo IP/SĐT\n\n**Swagger UI:** `/api/docs` · **OpenAPI JSON:** `/api/docs-json`",
    "version": "1.0.0",
    "contact": {
      "name": "FPT Telecom BE",
      "url": "",
      "email": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "accessToken từ login/register/refresh"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "Refresh JWT — tự gửi khi credentials: true"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "duong79",
            "description": "Tên đăng nhập (username)"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "SecretPass1"
          },
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn A"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "fullName"
        ]
      },
      "AuthUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "defaultAddress": {
            "type": "object",
            "nullable": true
          },
          "role": {
            "type": "string",
            "example": "CUSTOMER"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "fullName",
          "role"
        ]
      },
      "AdminProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "example": "ADMIN"
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token — gửi header Authorization: Bearer <token>"
          },
          "role": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "ADMIN"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          },
          "admin": {
            "$ref": "#/components/schemas/AdminProfileResponseDto"
          },
          "kind": {
            "type": "string",
            "example": "customer"
          }
        },
        "required": [
          "accessToken",
          "role"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Số điện thoại không hợp lệ"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "UnifiedLoginDto": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string",
            "example": "admin1",
            "description": "Admin: email đăng nhập. Khách: username."
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "account",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "duong79"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "role"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/AuthUserResponseDto"
          }
        },
        "required": [
          "role"
        ]
      },
...
      "CreateLeadDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "0912345678"
          },
          "installAddress": {
            "type": "string"
          },
          "packageId": {
            "type": "string",
            "description": "Nếu có, phải là gói đang active để snapshot"
          }
        },
        "required": [
          "fullName",
          "phone",
          "installAddress"
        ]
      },
      "PackageSnapshotDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "object",
            "nullable": true
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "type"
        ]
      },
      "LeadCreateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "NEW"
          },
          "createdAt": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "installAddress": {
            "type": "string"
          },
          "packageId": {
            "type": "object",
            "nullable": true
          },
          "packageSnapshot": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/PackageSnapshotDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "createdAt",
          "fullName",
          "phone",
          "installAddress"
        ]
      },
      "LeadPublicItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "installAddress": {
            "type": "string"
          },
          "packageId": {
            "type": "object",
            "nullable": true
          },
          "packageSnapshot": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/PackageSnapshotDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "createdAt",
          "fullName",
          "phone",
          "installAddress"
        ]
      },
      "LeadHistoryResponseDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "0912345678"
          },
          "total": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeadPublicItemDto"
            }
          }
        },
        "required": [
          "phone",
          "total",
          "items"
        ]
      },
...
      "AdminLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email hoặc username admin (vd. admin1, admin@example.com)",
            "example": "admin1"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "PatchAdminLeadDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "CONTACTED",
              "CONVERTED",
              "CANCELLED"
            ]
          },
          "adminNote": {
            "type": "string"
          }
        }
      },
...
    }
  }
}